requirements.txt
flask==3.0.2
python-docx==1.1.0
gunicorn==21.2.0